使用ffmpeg命令行进行mp4文件的合成

+++ app/src/main/java/com/hzl/smallvideo/manager/camera/CaptureButton.java
@@ -336,7 +336,7 @@ public class CaptureButton extends View {
             });
             record_anim.setInterpolator(null);
             //设置最大录制时间为15秒
-            record_anim.setDuration(RecordManager.RECORD_TIME * 1000);
+            record_anim.setDuration((long) (RecordManager.RECORD_TIME * 1000));
             record_anim.start();
         }
     }
